requirements.txt
# git+git://github.com/metashare/django.git@233117ea8786c711ba9ad34115e61077f4be65d6
# the commit above has only one modification compared to Django 1.4.19:
# extended permission name from 50 to 250 chars (django.contrib.auth.models.Permission.name)
Django==1.4.19
django-analytical==0.22.0
git+git://github.com/metashare/django-haystack.git@ea5dbd570f847ceb39ff741d9a27c77df09c9e3d
#django-haystack==2.0.0
# git+git://github.com/metashare/django-jenkins.git@e84f76ee7c940f35f761ced65ed7c97aa00b4acb
django-jenkins==0.13.0
# git+git://github.com/metashare/django-kronos.git@f22f638120920b722edd7619d6a037289edc678e
django-kronos==0.3
git+git://github.com/metashare/django-selectable.git@d8381086e8307257259a6cedf01d7fb2cbd34452
# django-selectable==0.5.2
django-selenium==0.9.7
djangosaml2==0.10.0
flup==1.0.2
httplib2==0.9.1
psycopg2==2.6.1
pycountry==1.12
pygeoip==0.3.2
pyoai==2.4.4
pysaml2==0.4.3
pysolr==2.1.0-beta
python-dateutil==2.4.2
requests==2.7.0
selenium==2.46.0
six==1.9.0
Unidecode==0.04.18
wsgiref==0.1.2
gunicorn
